Refactor TaskScheduleMode
