+++ app/src/main/java/com/yaroslavgamayunov/toodoo/data/mappers/TaskEntityMapper.kt
@@ -2,8 +2,9 @@ package com.yaroslavgamayunov.toodoo.data.mappers
 
 import com.yaroslavgamayunov.toodoo.data.api.TaskApiEntity
 import com.yaroslavgamayunov.toodoo.data.db.TaskRoomEntity
-import com.yaroslavgamayunov.toodoo.data.model.TaskScheduleMode
 import com.yaroslavgamayunov.toodoo.domain.entities.Task
+import com.yaroslavgamayunov.toodoo.domain.entities.TaskScheduleMode
+import com.yaroslavgamayunov.toodoo.util.taskScheduleMode
 import java.time.Instant
 import java.time.ZoneId
 import java.time.ZonedDateTime
@@ -17,7 +18,6 @@ fun Task.toTaskRoomEntity(
         description = description,
         isCompleted = isCompleted,
         deadline = deadline.toInstant(),
-        scheduleMode = scheduleMode,
         priority = priority,
         createdAt = createdAt,
         updatedAt = updatedAt
@@ -28,7 +28,7 @@ fun TaskRoomEntity.toTask() = Task(
     description = description,
     isCompleted = isCompleted,
     deadline = deadline.atZone(ZoneId.systemDefault()),
-    scheduleMode = scheduleMode,
+    scheduleMode = deadline.taskScheduleMode(),
     priority = priority
 )
 
@@ -47,6 +47,6 @@ fun TaskApiEntity.toTask() = Task(
     description = description,
     isCompleted = isCompleted,
     deadline = ZonedDateTime.ofInstant(Instant.ofEpochSecond(deadline), ZoneId.systemDefault()),
-    scheduleMode = TaskScheduleMode.ExactTime, // TODO: Remove
+    scheduleMode = Instant.ofEpochSecond(deadline).taskScheduleMode(),
     priority = priority
 )
