--- app/src/main/java/com/yaroslavgamayunov/toodoo/data/model/TaskScheduleMode.kt
@@ -1,7 +1,0 @@
-package com.yaroslavgamayunov.toodoo.data.model
-
-enum class TaskScheduleMode(val id: Int) {
-    ExactTime(0),
-    NotExactTime(1),
-    Unspecified(2)
-}
